Refactoring code to avoid duplication of functionality for IconLookup
